+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Info endpoint (/) for API v1
 - Add new datasets using the command line
 - Update existing datasets using the command line
+- Delete a dataset using the command line